chore: refactor DottetProgressBar and Employee
